+++ exportplan/views.py
@@ -2,9 +2,11 @@ from datetime import datetime
 import json
 import sentry_sdk
 
+from django.http import Http404
 from django.views.generic import TemplateView, FormView
 from django.utils.functional import cached_property
 from django.shortcuts import redirect
+from django.urls import reverse_lazy
 
 from rest_framework import generics
 from rest_framework.permissions import IsAuthenticated
@@ -15,33 +17,19 @@ from requests.exceptions import RequestException
 from directory_constants.choices import INDUSTRIES
 from directory_api_client.client import api_client
 from exportplan import data, helpers, serializers, forms
+from core.helpers import CountryDemographics
 
 
-class BaseExportPlanView(TemplateView):
+class ExportPlanMixin:
+
+    def dispatch(self, request, *args, **kwargs):
+        if self.slug not in data.SECTION_SLUGS:
+            raise Http404()
+        return super().dispatch(request, *args, **kwargs)
 
     @cached_property
     def export_plan(self):
         return helpers.get_or_create_export_plan(self.request.user)
-
-    def get_context_data(self, *args, **kwargs):
-        industries = [name for id, name in INDUSTRIES]
-        country_choices = [{'value': key, 'label': label} for key, label in helpers.get_madb_country_list()]
-        return super().get_context_data(
-            next_section=self.next_section,
-            sections=data.SECTION_TITLES,
-            sectors=json.dumps(industries),
-            country_choices=json.dumps(country_choices),
-            *args, **kwargs)
-
-
-class ExportPlanSectionView(BaseExportPlanView):
-
-    @property
-    def slug(self, **kwargs):
-        return self.kwargs['slug']
-
-    def get_template_names(self, **kwargs):
-        return [f'exportplan/sections/{self.slug}.html']
 
     @property
     def next_section(self):
@@ -54,9 +42,52 @@ class ExportPlanSectionView(BaseExportPlanView):
             'url': data.SECTION_URLS[index + 1],
         }
 
+    def get_context_data(self, **kwargs):
+        industries = [name for _, name in INDUSTRIES]
+        country_choices = [{'value': key, 'label': label} for key, label in helpers.get_madb_country_list()]
+        return super().get_context_data(
+            next_section=self.next_section,
+            sections=data.SECTION_TITLES,
+            export_plan=self.export_plan,
+            sectors=json.dumps(industries),
+            country_choices=json.dumps(country_choices),
+            **kwargs
+        )
+
+
+class ExportPlanSectionView(ExportPlanMixin, TemplateView):
+    @property
+    def slug(self, **kwargs):
+        return self.kwargs['slug']
+
+    def get_template_names(self, **kwargs):
+        return [f'exportplan/sections/{self.slug}.html']
+
+
+class ExportPlanMarketingApproachView(ExportPlanMixin, FormView):
+    form_class = forms.CountryDemographicsForm
+    template_name = 'exportplan/sections/marketing-approach.html'
+    slug = 'marketing-approach'
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+        data = self.request.GET or {}
+        if data:
+            kwargs['data'] = data
+        return kwargs
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        form = self.get_form()
+        if form.is_valid():
+            country = CountryDemographics(form.cleaned_data['name'])
+            context['country'] = country
+            context['age_range'] = country.filter_age_range(form.cleaned_data['age_range'])
+            context['united_kingdom'] = CountryDemographics('United Kingdom')
+        return context
+
 
 class ExportPlanTargetMarketsView(ExportPlanSectionView):
-    slug = 'target-markets'
     template_name = 'exportplan/sections/target-markets.html'
 
     def get_context_data(self, **kwargs):
@@ -68,6 +99,55 @@ class ExportPlanTargetMarketsView(ExportPlanSectionView):
         )
 
 
+class FormContextMixin:
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(*args, **kwargs)
+
+        field_names = list(self.form_class.base_fields.keys())
+
+        field_labels = [field.label for field in self.form_class.base_fields.values()]
+
+        field_placeholders = [
+            field.widget.attrs.get('placeholder', '') for field in self.form_class.base_fields.values()
+        ]
+
+        form_fields = [
+            {'name': name, 'label': label, 'placeholder': placeholder}
+            for name, label, placeholder in zip(field_names, field_labels, field_placeholders)
+        ]
+
+        context['form_initial'] = json.dumps(context['form'].initial)
+        context['form_fields'] = json.dumps(form_fields)
+
+        return context
+
+
+class ExportPlanBrandAndProductView(FormContextMixin, ExportPlanSectionView, FormView):
+    form_class = forms.ExportPlanBrandAndProductForm
+    success_url = reverse_lazy('exportplan:brand-and-product')
+
+    def form_valid(self, form):
+        helpers.update_exportplan(
+            sso_session_id=self.request.user.session_id,
+            id=self.export_plan['pk'],
+            data={'brand_product_details': form.cleaned_data}
+        )
+        return super().form_valid(form)
+
+
+class ExportPlanBusinessObjectivesView(FormContextMixin, ExportPlanSectionView, FormView):
+    form_class = forms.ExportPlanBusinessObjectivesForm
+    success_url = reverse_lazy('exportplan:objectives')
+
+    def form_valid(self, form):
+        helpers.update_exportplan(
+            sso_session_id=self.request.user.session_id,
+            id=self.export_plan['pk'],
+            data=form.cleaned_data
+        )
+        return super().form_valid(form)
+
+
 class UpdateExportPlanAPIView(generics.GenericAPIView):
 
     serializer_class = serializers.ExportPlanSerializer
@@ -75,14 +155,17 @@ class UpdateExportPlanAPIView(generics.GenericAPIView):
 
     def post(self, request, *args, **kwargs):
         serializer = self.serializer_class(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        export_plan = helpers.get_or_create_export_plan(self.request.user)
-        helpers.update_exportplan(
-            sso_session_id=self.request.user.session_id,
-            id=export_plan['pk'],
-            data=serializer.validated_data
-        )
-        return Response({})
+
+        if serializer.is_valid(raise_exception=True):
+            export_plan = helpers.get_or_create_export_plan(self.request.user)
+            helpers.update_exportplan(
+                sso_session_id=self.request.user.session_id,
+                id=export_plan['pk'],
+                data=serializer.validated_data
+            )
+            return Response(serializer.validated_data)
+
+        return Response(serializer.errors)
 
 
 class BaseFormView(FormView):