+++ react-components/src/bundle.js
@@ -12,6 +12,8 @@ import LoginModal from '@src/views/LoginModal/Modal'
 import MarketSelectNavbar from '@src/views/MarketSelectNavbar/Container'
 import Tour from '@src/views/Tour/Container'
 import { createTargetMarketCountries } from '@src/components/TargetMarketCountries'
+import { createBrandAndProductForm } from '@src/components/BrandAndProduct'
+import { createBusinessRationale } from '@src/components/BusinessRationale'
 import LearnIntroduction from '@src/views/LearnIntroduction/LearnIntroduction'
 import { STEP_CREDENTIALS, STEP_VERIFICATION_CODE } from '@src/views/SignupModal/Component/'
 import Services from '@src/Services'
@@ -22,6 +24,8 @@ export default {
   Countries,
   CountriesModal,
   createTargetMarketCountries,
+  createBusinessRationale,
+  createBrandAndProductForm,
   IndustriesModal,
   LearnIntroduction,
   ProductLookup,