+++ CHANGELOG.md
@@ -10,8 +10,13 @@
 - MVP-437 - Show product in lesson page nav bar
 - No ticket - v3 ci pipeline manifest.yml file fix
 - MVP-443 - Use non-GREAT stlye 404 and 500 pages
+- MVP-579 - create export plan cms page
+- GP2-67 - fix backbutton upload logo
 
 ### Implemented enhancements
+- MVP-392 - Add remove sector functionality
+- GP2-115 - Add objectives business rationale section
+- MVP-452 - Add brand and product page with autosave function
 - MVP-317 - Add dummy pages for all export plan sections
 - MVP-306 - Added sector chooser component
 - MVP-299 - Finalised styling for export plan sections
@@ -98,3 +103,6 @@
 - no ticket - Handle dit-cf-security munging cookies
 - no ticket - fix logo
 - no ticket - fix locust's breaking changes
+- MVP-563 - Add time to read
+- MVP-564 - Collect country demographics information from spreadsheets
+- GV2-104 - GP2-121 - Tags and content modules